Resolves four open issues across SEO, wallet security, and accessibility for the ProofOfHeart frontend
