+++ src/app/[locale]/causes/[id]/page.tsx
@@ -1,9 +1,16 @@
-"use client";
-
-import { useParams } from "next/navigation";
+import { buildAlternates } from "@/lib/seo";
 import CauseDetailClient from "./CauseDetailClient";
 
-export default function Page() {
-  const { id } = useParams<{ id: string }>();
+type Props = { params: Promise<{ locale: string; id: string }> };
+
+export async function generateMetadata({ params }: Props) {
+  const { id } = await params;
+  return {
+    alternates: buildAlternates(`/causes/${id}`),
+  };
+}
+
+export default async function Page({ params }: Props) {
+  const { id } = await params;
   return <CauseDetailClient id={id} />;
 }
