+++ src/lib/seo.ts
@@ -1,0 +1,12 @@
+import { routing } from "@/i18n/routing";
+
+const SITE_URL = process.env.NEXT_PUBLIC_SITE_URL ?? "https://proofofheart.xyz";
+
+export function buildAlternates(path: string) {
+  const languages: Record<string, string> = {};
+  for (const locale of routing.locales) {
+    languages[locale] = `${SITE_URL}/${locale}${path}`;
+  }
+  languages["x-default"] = `${SITE_URL}/${routing.defaultLocale}${path}`;
+  return { languages };
+}
